ImageProcessor/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pipreqs==0.4.10
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
yarg==0.1.9
